requirements.txt
streamlit
pandas
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
plotly
